Variavel com ponto flutuante

+++ src/Main.java
@@ -5,11 +5,11 @@ public class Main {
 	public static void main(String[] args) {
 		Locale.setDefault(Locale.US);
 		
-		double b, B, h, area;
-		b = 6.0;
-		B = 8.0;
-		h = 5.0;
-		area = (b + B) / 2.0 * h;
+		float b, B, h, area;
+		b = 6f;
+		B = 8f;
+		h = 5f;
+		area = (b + B) / 2f * h;
 		System.out.println(area);
 		
 	}
